ball의 상위 클래스 Circle 작성
