+++ yongjae/src/main/java/com/nhnacademy/Ball.java
@@ -1,30 +1,16 @@
 package com.nhnacademy;
 
-import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
-public class Ball {
-    private double x; // 공의 현재 x 좌표
-    private double y; // 공의 현재 y 좌표
-    private double radius; // 공의 반지름
+public class Ball extends Circle {
     private double dx; // 공의 x축 속도 (단위: 픽셀/프레임)
     private double dy; // 공의 y축 속도 (단위: 픽셀/프레임)
-    private Color color; // 공의 색상
 
     // 생성자
     public Ball(double x, double y, double radius, double dx, double dy, Color color) {
-        this.x = x;
-        this.y = y;
-        this.radius = radius;
+        super(x, y, radius, color);
         this.dx = dx;
         this.dy = dy;
-        this.color = color;
-    }
-
-    // 공을 그리는 메서드
-    public void draw(GraphicsContext gc) {
-        gc.setFill(color);
-        gc.fillOval(x - radius, y - radius, radius * 2, radius * 2); // 중심을 기준으로 원 그리기
     }
 
     // 공의 위치를 업데이트하는 메서드
@@ -45,26 +31,11 @@ public class Ball {
         }
     }
 
-    // Getter와 Setter (필요 시 사용)
-    public double getX() {
-        return x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public double getRadius() {
-        return radius;
-    }
-
     public double getDx() {
         return dx;
     }
 
-    public double getDy() {
-        return dy;
-    }
+    public double getDy() { return dy; }
 
     public void setDx(double dx) {
         this.dx = dx;
